+++ addons/armor/CfgVehicles.hpp
@@ -76,7 +76,7 @@ class CfgVehicles {
     #include "configs\Units_P2_Tanker.hpp"
     #include "configs\Units_ARC.hpp"
     #include "configs\Units_ARF.hpp"
-    #include "configs\Units_Commando.hpp"
+    #include "configs\Units_cloneCommando.hpp"
     #include "configs\Units_Jedi.hpp"
 
     class CLASS(backpack_base);
